Reorganize Numba functions for mag simulations
